refactor(web): query state 네이밍과 result mapper 구조 분리개선
