+++ apps/web/src/features/space-search/model/query-keys.ts
@@ -1,7 +1,7 @@
-import type { SearchResultsQueryState } from "./search-params";
+import type { SearchRequestQueryState } from "./types";
 
 export const searchQueryKeys = {
   all: ["search-results"] as const,
-  list: (queryState: SearchResultsQueryState, isAuthenticated: boolean) =>
+  list: (queryState: SearchRequestQueryState, isAuthenticated: boolean) =>
     [...searchQueryKeys.all, queryState, { isAuthenticated }] as const,
 };
